% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.co.mapping.R
\name{create.co.mapping}
\alias{create.co.mapping}
\title{create.co.mapping}
\usage{
create.co.mapping(co)
}
\arguments{
\item{co}{Cardiac output waveform.}
}
\value{
Function mapping t to cardiac output
}
\description{
Given a 125 Hz CO waveform, returns a function mapping time to cardiac output.
}
