% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_co_parlikar.R
\name{estimate.co.parlikar}
\alias{estimate.co.parlikar}
\title{ESTIMATE_CO_PARLIKAR}
\usage{
estimate.co.parlikar(abp, feat, onsets, window.radius)
}
\arguments{
\item{abp}{Arterial blood pressure waveform (125 Hz, mmHg)}

\item{feat}{Features computed using abpfeature()}

\item{onsets}{Beat onsets computed using wabp()}

\item{window.radius}{Window under which we compute least-squares estimate of the time constant}
}
\value{
Vector of uncalibrated beat-to-beat estimates of cardiac output
}
\description{
ESTIMATE_CO_PARLIKAR
}
