% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsqi.R
\name{jsqi}
\alias{jsqi}
\title{JSQI}
\usage{
jsqi(features, onset, abp)
}
\arguments{
\item{features}{Features computed using abpfeature()}

\item{onset}{Onsets computed using wabp()}

\item{abp}{ABP waveform (125Hz sampled, mmHg units of measure)}
}
\value{
The following signal quality metrics:
1. Logical OR of cols 2 through 10
2. P not physiologic 
3. MAP not physiologic
4. HR not physiologic
5. PP not physiologic
6. abnormal Psys
7. abnormal Pdias
8. abnormal period
9. abnormal P(onset)
10. noisy beat
}
\description{
Signal quality index calculator.
Original MATLAB code by James Sun
Ported to R by Ran Liu on June 18, 2019
}
