% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liljestrand.estimate.co.waveform.R
\name{liljestrand.estimate.co.waveform}
\alias{liljestrand.estimate.co.waveform}
\title{liljestrand.estimate.co.waveform}
\usage{
liljestrand.estimate.co.waveform(abp, i0)
}
\arguments{
\item{abp}{Arterial blood pressure waveform}

\item{i0}{Average value of cardiac output over interval.}
}
\value{
Cardiac output waveform estimated using Liljestrand estimator.
}
\description{
Returns a function mapping t to cardiac output.
}
