% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wabp.R
\name{wabp}
\alias{wabp}
\title{WABP}
\usage{
wabp(abp)
}
\arguments{
\item{abp}{125 Hz arterial blood pressure waveform data in physiological units.}
}
\value{
Index of the onset of each heartbeat.
}
\description{
A port of Sun et al's ABP waveform onset detector.
WABP(ABP) obtains the onset time (in samples) of each beat in the ABP waveform.
}
\details{
Original code by James Sun on Nov 19, 2005.
Ported to R by Ran Liu on June 24, 2018.
}
