% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windkessel.3.element.R
\name{windkessel.three.element}
\alias{windkessel.three.element}
\title{windkessel.three.element}
\usage{
windkessel.three.element(T, N, r1, r2, c, p0, cardiac.output)
}
\arguments{
\item{T}{end time}

\item{N}{number of discrete time steps for the interval [0,T]}

\item{r1}{Resistance 1}

\item{r2}{Resistance 2}

\item{c}{Capacitance}

\item{p0}{Initial value of pressure}

\item{i}{Function that maps time to cardiac output (mL/s)}
}
\value{
Waveform of ABP over time
}
\description{
Generates data according to the 3-element windkessel model
}
